+++ packages/ui/resources/views/partials/editor-widget-scripts.blade.php
@@ -1,2 +1,3 @@
 <script src="{{ \Velm\Ui\UiAssets::richTextScriptHref() }}" data-navigate-track></script>
 <script src="{{ \Velm\Ui\UiAssets::codeEditorScriptHref() }}" data-navigate-track></script>
+<script src="{{ \Velm\Ui\UiAssets::codeDisplayScriptHref() }}" data-navigate-track></script>